Add AppointmentList class and listAppointment function
