+++ src/main/java/AppointmentList.java
@@ -1,0 +1,15 @@
+import java.util.ArrayList;
+
+public class AppointmentList {
+    static ArrayList<Appointment> appointments = new ArrayList<>();
+
+    public static void listAppointment() {
+        for (Appointment appointment : appointments) {
+            System.out.print(appointment.appointmentId + " ");
+            System.out.print(appointment.petName + " ");
+            // System.out.print(appointment.service.name + " ");
+            System.out.print(appointment.getAppointmentDate() + " ");
+            System.out.println(appointment.getAppointmentStatus());
+        }
+    }
+}
